// src/app/modules/addToCart/addToCart.interface.ts
import { Schema } from "mongoose";

export type TAddToCart = {
	userId: Schema.Types.ObjectId; // Reference to the user who added the product to their cart
	productId: Schema.Types.ObjectId; // Reference to the product added to the cart
	quantity: number; // Quantity of the product in the cart
	addedAt: Date; // Date when the product was added to the cart
	isPurchased: boolean; // Indicates if the product has been purchased
};